fix: build, add api ref layout
